requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
psycopg[binary]==3.2.12
boto3==1.35.0
requests==2.32.3
openai==1.52.2
httpx==0.27.2  # pinned for compatibility with openai 1.52.2
bcrypt==4.2.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-cov==5.0.0
pytest-xdist==3.6.0
python-docx==1.2.0
pdfplumber==0.11.8
pandas==2.3.3
better-profanity==0.7.0

